+++ NINA.Plugin.Web/NINA.Plugin.Web/ImageSaveWatcher.cs
@@ -1,26 +1,23 @@
 ﻿using NINA.Core.Utility;
 using NINA.Equipment.Model;
+using NINA.Profile.Interfaces;
 using NINA.WPF.Base.Interfaces.Mediator;
 using System;
-using System.ComponentModel;
 using Web.NINAPlugin.History;
 
 namespace Web.NINAPlugin {
 
     public class ImageSaveWatcher {
         private IImageSaveMediator imageSaveMediator;
-        private bool WebPluginEnabled;
-        private bool NonLightsEnabled;
+        private IPluginOptionsAccessor pluginSettings;
         private bool running = false;
 
         private SessionHistoryManager sessionHistoryManager;
         private string sessionHome;
 
-        public ImageSaveWatcher(IImageSaveMediator imageSaveMediator) {
+        public ImageSaveWatcher(IImageSaveMediator imageSaveMediator, IPluginOptionsAccessor pluginSettings) {
             this.imageSaveMediator = imageSaveMediator;
-            WebPluginEnabled = Properties.Settings.Default.WebPluginEnabled;
-            NonLightsEnabled = Properties.Settings.Default.NonLights;
-            Properties.Settings.Default.PropertyChanged += SettingsChanged;
+            this.pluginSettings = pluginSettings;
 
             sessionHistoryManager = new SessionHistoryManager();
         }
@@ -49,8 +46,8 @@ namespace Web.NINAPlugin {
         }
 
         private void ImageSaveMeditator_ImageSaved(object sender, ImageSavedEventArgs msg) {
-            if (!WebPluginEnabled) {
-                Logger.Debug("web plugin not enabled");
+            if (!isPluginActive()) {
+                Logger.Debug("web plugin not active");
                 return;
             }
 
@@ -77,7 +74,7 @@ namespace Web.NINAPlugin {
                 return false;
             }
 
-            if (!NonLightsEnabled) {
+            if (!isNonLightsEnabled()) {
                 return imageType == CaptureSequence.ImageTypes.LIGHT;
             }
 
@@ -113,16 +110,13 @@ namespace Web.NINAPlugin {
             return msg.MetaData.Target.Name;
         }
 
-        private void SettingsChanged(object sender, PropertyChangedEventArgs e) {
-            switch (e.PropertyName) {
-                case "WebPluginEnabled":
-                    WebPluginEnabled = Properties.Settings.Default.WebPluginEnabled;
-                    break;
+        private bool isPluginActive() {
+            string state = pluginSettings.GetValueString(nameof(WebPlugin.WebPluginState), WebPlugin.WebPluginStateDefault);
+            return state.Equals(WebPlugin.WebPluginStateON) || state.Equals(WebPlugin.WebPluginStateSHARE);
+        }
 
-                case "NonLights":
-                    NonLightsEnabled = Properties.Settings.Default.NonLights;
-                    break;
-            }
+        private bool isNonLightsEnabled() {
+            return pluginSettings.GetValueBoolean(nameof(WebPlugin.NonLights), WebPlugin.NonLightsDefault);
         }
     }
 }