+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Web Session History Viewer
 
+## 1.1.2.0 - 2022-05-04
+* Plugin options are now saved per NINA profile
+* Now supports running another instance of NINA in 'share' mode
+* Web server will now automatically redirect root path requests
+* NINA profile name now displayed in the Web client
+* Dates are now ISO formatted
+* Bug fix: fixed issue with domain of Y axes in quality chart
+
 ## 1.1.0.1 - 2022-04-20
 * Bug fix: fixed problems with loading of problematic session histories
 * Bug fix: better evaluation and handling of AF trend lines